feat: use custom overrides to handle interrupts and only count

+++ src/main/java/org/collection/fm/analyses/IndependentCyclomaticComplexity.java
@@ -24,7 +24,7 @@ public class IndependentCyclomaticComplexity implements IFMAnalysis {
 
     @Override
     public String getResult(IFeatureModel featureModel, FeatureModelFormula formula, int timeout, Path solverRelativePath) {
-        return Integer.toString(ConnectivityGraph.getNumberOfIndependentCycles(formula));
+        return Long.toString(ConnectivityGraph.getNumberOfIndependentCycles(formula));
     }
 
     @Override
